chore: move all the diffing APIs in hooks.ts
